Revert "Dummy pr"

+++ database/migrations/2026_07_29_214404_create_pull_requests_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
         Schema::create('pull_requests', function (Blueprint $table) {
             $table->id();
             $table->timestamps();
-            $table->bigInteger('github_id')->unique();
+            $table->integer('github_id')->unique();
             $table->foreignId('github_repository_id')->constrained('github_repositories');
             $table->integer('number');
             $table->string('title');
